% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_example_doc.R
\docType{data}
\name{city_example}
\alias{city_example}
\title{City example}
\description{
It contains a urban model as an example to use the package.
It is specifically the urban model for Sant Narcís, a neighbourhood of Girona (Catalonia)
}
\author{
Josep Pueyo-Ros
}
