% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_functions_doc.R
\docType{data}
\name{city_functions}
\alias{city_functions}
\title{City Functions}
\description{
It contains information about the functions belonging to green infrastructure and urban agriculture
in the city_example dataset. This information can be used as params in the indicators, in some cases these
are already the default values.
}
\author{
Josep Pueyo-Ros
}
