% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_land_uses_doc.R
\docType{data}
\name{city_land_uses}
\alias{city_land_uses}
\title{City Land Uses}
\description{
It contains information about the functions belonging to green infrastructure and urban agriculture
in the city_example dataset. This information can be used as params in the indicators, in some cases these
are already the default values.
}
\author{
Josep Pueyo-Ros
}
