% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_jobs.R
\name{edible_jobs}
\alias{edible_jobs}
\title{The jobs created by urban agriculture in your city}
\usage{
edible_jobs(
  x,
  jobs = c(0.000163, 0.022),
  edible = c("Commercial gardens", "Hydroponic rooftops"),
  area_field = "flat_area"
)
}
\arguments{
\item{x}{An 'sf' object with the urban model of your city.}

\item{edible}{The categories in 'Functions' that represent commercial edible gardens.}

\item{area_field}{The field to be used as the area of each feature. If NULL, the area is calculated with
sf::st_area()}
}
\description{
This indicator estimates the number of full-time jobs created by commercial urban agriculture
initiatives in your city. It uses a range of jobs per square meter to create the median and the
confidence interval of the number of jobs by simulating a random uniform distribution of 1000 values
within the provided range.
}
