% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_volunteers.R
\name{edible_volunteers}
\alias{edible_volunteers}
\title{The number of volunteers involved in urban agriculture in your city}
\usage{
edible_volunteers(
  x,
  volunteers = c(0.000163, 0.022),
  edible = c("Community gardens", "Rooftop gardens"),
  area_field = "flat_area"
)
}
\arguments{
\item{x}{An 'sf' object with the urban model of your city.}

\item{edible}{The categories in 'Functions' that represent community edible gardens.}

\item{area_field}{The field to be used as the area of each feature. If NULL, the area is calculated with
sf::st_area()}
}
\description{
This indicator estimates the number of volunteers potentially involved in community urban agriculture
initiatives in your city. It uses a range of volunteers per square meter to create the median and the
confidence interval of the number of volunteers by simulating a random uniform distribution of 1000 values
within the provided range. The default range came from required work hours in urban agriculture assessed
in scientific literature, assuming that a volunteers dedicates a 10% of a full-time job.
}
