% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_green_capita.R
\name{green_capita}
\alias{green_capita}
\title{Urban green per capita}
\usage{
green_capita(
  x,
  green_categories = NULL,
  inhabitants = NULL,
  neighbourhoods = NULL,
  inh_field = "inhabitants",
  private = FALSE
)
}
\arguments{
\item{x}{An 'sf' object with the urban model of your city and a 'Function' field with categories of urban features.}

\item{green_categories}{The categories that are considered as urban green. If NULL, categories of 'get_categories()'
are considered.}

\item{inhabitants}{A value representing the inhabitants in the city.}

\item{inh_field}{(optional) The field in 'neighbourhoods' indicating the inhabitants in each neighbourhood.}

\item{private}{If FALSE (default), only public areas are considered in the indicator.}

\item{neighbourboods}{(optional) An 'sf' object with polygons representing the neighbourhoods in the city.}
}
\description{
This indicators calculates the amount of green per capita in the city. This may include private green
such as gardens and crops or exclude them.
}
\details{
If neighbourhoods are provided the 'inh_field' must also be provided. Then, 'inhabitants' is ignored.
}
