% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods_doc.R
\docType{data}
\name{neighbourhoods}
\alias{neighbourhoods}
\title{Neighbourhoods polygons of city_example dataset}
\description{
It contains an 'sf' object with the polygons of the neighbourhoods of the city_example. It is used
in some indicators when they are disaggregated by neighbourhoods.
}
\author{
Josep Pueyo-Ros
}
