% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourhoods_doc.R
\docType{data}
\name{neighbourhoods_example}
\alias{neighbourhoods_example}
\title{Neighbourhoods of City Example}
\description{
It contains a GIS layer with the neighbourhoods of city_example and their population.
It is used to run 'green_capita' indicator.
}
\author{
Josep Pueyo-Ros
}
