% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_NO2.R
\name{no2_seq}
\alias{no2_seq}
\title{Sequestration of nitrogen dioxide}
\usage{
no2_seq(x, green_df = NULL, return_raster = F, verbose = F)
}
\arguments{
\item{x}{An 'sf' object with the urban model of your city and a 'Function' column with categories of
urban features.}

\item{green_df}{A dataframe of categories that are considered as urban green with two columns. 'functions'
with the names of 'Function' in 'x' to be considered as green and a 'no2_seq' columns with NO2
sequestration capacity of each function. If NULL, categories and values of 'city_functions' dataset are
considered.}
}
\value{
If 'return_raster is TRUE, a 'stars' object is returned with NO2 sequestration as attribute. If 'verbose'
is TRUE, a vector with the values of NO2 sequestration. If previous both are FALSE, the summary of values is returned.
}
\description{
This indicator returns the amount of NO2 that is sequestered by urban green.
}
