% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_runoff.R
\name{runoff_prev}
\alias{runoff_prev}
\title{Runoff prevention}
\usage{
runoff_prev(
  x,
  runoff_df = NULL,
  rain = 85,
  floors_field = "floors",
  harvest_dist = 10,
  tank_size = c(0, 45)
)
}
\arguments{
\item{x}{An 'sf' object with the urban model of your city and a 'Function' column with categories
of urban features.}

\item{runoff_df}{A dataframe of categories that are considered impervious area with five columns. 'functions'
with the names of 'Function' in 'x' to be considered as impervious; two infiltration columns 'infiltration1'
and 'infiltration2 with the range of infiltration capacity (mm in 24 hours) of that function;
and two water storage capacity columns 'water_storage1' and 'water_storage2' with the range of water
storage capacity of that function. If NULL, categories and values of 'city_functions' dataset
are considered.}

\item{rain}{The amount of 24h-rain to be simulated, default is 100 mm.}

\item{floors_field}{The column in 'x' containing the number of floors of each building. Zero is considered
unbuilt areas like gardens or streets. It is used to calculate rainwater haversting area, since only
upper surface are considered. Missing values are considered as zero.}

\item{harvest_dist}{Maximum distance (in meters) of buildings where to harvest rainwater}

\item{tank_size}{A two-length vector with the range of tank size possibilities (in l/m2).}
}
\value{
It returns a named vector with values of percentage of rainfall infiltred or captured,
total rainfall, total infiltration and total rainharvest (all of them in m3)
}
\description{
The indicator calculates the runoff prevention considering a rain event, the infiltration
capacity and the rain harvesting and storage capacity.
government.
}
\author{
Josep Pueyo-Ros
}
