% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf_doc.R
\docType{data}
\name{SVF}
\alias{SVF}
\title{Sky View Factor for City Example}
\description{
It contains a raster representing the sky view factor for 'city_example',
calculated using SAGA algorithm on a digital surface model of 3x3 from a LIDAR dataset.
}
\author{
Josep Pueyo-Ros
}
