% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_scenario.R, R/set_scenario_rationale.R
\name{set_scenario}
\alias{set_scenario}
\title{Set the scenario for your edible city}
\usage{
set_scenario(
  x,
  pGardens = 1,
  pVacant = 1,
  pRooftop = 1,
  perc_garden = c(0.02, 0.3),
  perc_vacant = c(0.52, 0.75),
  perc_rooftop = c(0.6, 0.62),
  min_area_garden = 10,
  min_area_vacant = 100,
  min_area_rooftop = 100,
  private_gardens_from = "Normal garden",
  vacant_from = "Vacant",
  rooftop_from = "Rooftop",
  pCommercial = 0,
  area_field = "flat_area"
)

set_scenario(
  x,
  pGardens = 1,
  pVacant = 1,
  pRooftop = 1,
  perc_garden = c(0.02, 0.3),
  perc_vacant = c(0.52, 0.75),
  perc_rooftop = c(0.6, 0.62),
  min_area_garden = 10,
  min_area_vacant = 100,
  min_area_rooftop = 100,
  private_gardens_from = "Normal garden",
  vacant_from = "Vacant",
  rooftop_from = "Rooftop",
  pCommercial = 0,
  area_field = "flat_area"
)
}
\arguments{
\item{x}{An 'sf' object with the urban model of your city and a 'Function' field with categories of urban features.}

\item{pGardens}{The proportion of private gardens (Function == 'Gardens')
that will become edible gardens [0-1].}

\item{pVacant}{The proportion of vacant plot (Function == 'Vacant') with 'area >= min_area_vacant'
that will become edible gardens [0-1].}

\item{pRooftop}{The proportion of rooftops (Function == 'Flat rooftop') with 'area >= min_area_rooftop'
that will become edible rooftops [0-1].}

\item{perc_garden}{The proportion in a range of surface in a garden that is occupied by edible plants [0-1].}

\item{perc_vacant}{The proportion in a range of surface in a vacant plot that is occupied by edible plants [0-1].}

\item{perc_rooftop}{The proportion in a range of surface in a rooftop that is occupied by edible plants [0-1].}

\item{min_area_garden}{The minimum area that a garden must have to become an edible garden.}

\item{min_area_vacant}{The minimum area that a vacant must have to become an community or commercial garden.}

\item{min_area_rooftop}{The minimum area that a flat rooftop must have to become an edible rooftop.}

\item{private_gardens_from}{The categories in 'Functions' potentially converted to edible private gardens}

\item{vacant_from}{The categories in 'Functions' potentially converted to community or commercial gardens}

\item{rooftop_from}{The categories in 'Functions' potentially converted to edible rooftop
(community raised beds or commercial hydroponic)}

\item{pCommercial}{The proportion of plots and rooftop that will be commercial. The rest will be community gardens
In rooftops it is equivalent to raised beds and hydroponic system respectively.}

\item{area_field}{The field to be used as the area of each feature. If NULL, the area is calculated with
sf::st_area()}
}
\value{
An 'sf' object as 'x' with the respective proportion of gardens ('Edible private garden'),
vacant plots ('Community plot garden', 'Commercial plot garden') and rooftop gardens ('Community rooftop garden',
'Commercial hydroponic rooftop')
labelled as edible gardens.

An 'sf' object as 'x' with the respective proportion of gardens ('Edible private garden'),
vacant plots ('Community plot garden', 'Commercial plot garden') and rooftop gardens ('Community rooftop garden',
'Commercial hydroponic rooftop')
labelled as edible gardens.
}
\description{
You can adjust different parameters to define different city scenarios.
The object must contain a field 'Function' which describes the function or type of each feature.

You can adjust different parameters to define different city scenarios.
The object must contain a field 'Function' which describes the function or type of each feature.
}
\details{
When pGardens is lower than 1, the gardens are selected randomly among gardens with an area larger than
'min_area_garden'. However, vacant plots and rooftops are selected from larger to smaller, assuming that the best
spots (i.e. larger) are occupied first. Likewise, when pCommercial > 0, commercial gardens and hydroponic rooftops are
settled in the larger features, assuming that commercial initiatives have the power to acquire the best spots.

When pGardens is lower than 1, the gardens are selected randomly among gardens with an area larger than
'min_area_garden'. However, vacant plots and rooftops are selected from larger to smaller, assuming that the best
spots (i.e. larger) are occupied first. Likewise, when pCommercial > 0, commercial gardens and hydroponic rooftops are
settled in the larger features, assuming that commercial initiatives have the power to acquire the best spots.
}
\author{
Josep Pueyo-Ros

Josep Pueyo-Ros
}
