% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_scenario.R
\name{set_scenario}
\alias{set_scenario}
\title{Dibuja el histograma de una variable superponiendo la densidad normal ajustada}
\usage{
set_scenario()
}
\arguments{
\item{x}{vector de datos cuyo histograma se va a calcular}

\item{dens}{valor lógico: TRUE=Superponer estimador de núcleo de la densidad}
}
\value{
el histograma con la densidad normal superpuesta
}
\description{
Función que dibuja el histograma de una variable x, superponiendo la densidad normal
ajustada. Si el usuario lo desea puede superponer también un estimador de núcleo de la densidad.
}
