% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_heat_island.R
\name{UHI}
\alias{UHI}
\title{Heat island effect}
\usage{
UHI(
  x,
  green_categories = NULL,
  pGreen = NULL,
  SVF,
  Qql = 6.11,
  Cair = 1007,
  Pair = 1.14,
  Tmax = 30.8,
  Tmin = 20,
  windspeed = 2.77,
  return_raster = F,
  verbose = F
)
}
\arguments{
\item{x}{An 'sf' object with the urban model of your city and a 'Function' column with categories of urban features.}

\item{green_categories}{A vector of categories that are considered as urban green. If NULL, categories of 'get_categories()'
are considered.}

\item{pGreen}{A vector of same length than green_categories with the percentage of green in each category.
If pGreen is NULL, green_categories are considered totally green.}

\item{SVF}{A 'stars' object representing sky view factor. It can be computed, e.g. with SAGA's
Sky View Factor algorithm and then loaded with stars::read_stars().}

\item{Qql}{A numerical value representing the average solar radiation in W/m2/hour.}

\item{Cair}{A numerical value representing the air heat capacity in J.}

\item{Pair}{A numerical value representing the air density in kg/m3.}

\item{Tmax}{Averaged maximum temperature in ºC.}

\item{Tmin}{Averaged minimum temperature in ºC.}

\item{return_raster}{If TRUE, the raster of UHI values is returned. Otherwise, a summary of raster values is returned.}

\item{verbose}{If TRUE, returns a vector with UHI value in each cell.}
}
\value{
A 'stars' object with values of UHI. Or a numerical vector or summary statistic for UHI values.
See params for more information on how to select each one.
}
\description{
The indicator calculates de urban heat island (UHI) using the DPRA guidelines of the Dutch
government.
}
\details{
DEFAULT values are the values for 'city_example' dataset in August (averaged values from 2011-2020)
}
